// src/api/models/cartResponse.ts
/**
 * Generated by orval v7.17.2 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { CartItemResponse } from './cartItemResponse';

export interface CartResponse {
  userId?: number;
  items?: CartItemResponse[];
  totalItems?: number;
  totalAmount?: number;
  selectedAmount?: number;
  selectedCount?: number;
}